requirements.txt
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
flask==3.1.0
streamlit>=1.35.0
pandas>=2.0.0
